Implement remove test
